requirements.txt
boto3==1.34.0
psycopg2-binary==2.9.9
pandas==2.1.4
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.16.0
google-generativeai==0.3.2
flask==3.0.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.13.1
webdriver-manager==4.0.1
lxml==4.9.3
scrapy==2.11.0
fastapi==0.104.1
uvicorn==0.24.0
serverless-wsgi==2.0.2
